Ping www.freedesktop.org instead of apt source
Instead of asking apt for the list of repositories and pinging one of
them to discover whether we are connected to the internet, let's ping
one fixed distribution agnostic host. www.freedosktop.org sounds like a
good site. Main reasoning behind this is to make network test working
regardless of distribution used. Probably could be done better, so
feel free to rewrite it if you have any better idea. For now, it will
help us to get further.